Add read_options and write_options support for LET's operations
Caller can tailor the read_options and write_options for LET's
operations by managing multiple table identifiers.  The function tid/2
returns a copy of a table's identifier with given LevelDB read and
write options.  If options are not undefined, the given read and write
options are used when the returned copy is passed as the first
argument to table operations.  Otherwise, the original read and write
options given at the time of calling new/2 are used.